obfuscate before serialization
