+++ agent/obfuscate.js
@@ -54,25 +54,6 @@ function validateRules (rules) {
   return !invalidReplacementDetected && !invalidRegexDetected
 }
 
-// traverses an object and applies a fn to property values of a certain type
-function applyFnToProps(obj, fn, type) {
-  if (!obj || typeof obj !== 'object') return obj
-  type = type || 'string'
-  return traverse(obj)
-  function traverse(obj) {
-    for (var property in obj) {
-      if (obj.hasOwnProperty(property)) {
-        if (typeof obj[property] === 'object') {
-          traverse(obj[property])
-        } else {
-          if (typeof obj[property] === type) obj[property] = fn(obj[property])
-        }
-      }
-    }
-    return obj
-  }
-}
-
   // applies all regex obfuscation rules to provided URL string and returns the result
 function obfuscateString (string) {
     // if string is empty string, null or not a string, return unmodified
@@ -91,7 +72,6 @@ function obfuscateString (string) {
 }
 
 module.exports = {
-  applyFnToProps: applyFnToProps,
   obfuscateString: obfuscateString,
   shouldObfuscate: shouldObfuscate,
   getRules: getRules,
